requirements.txt
babel==2.17.0
certifi==2025.4.26
charset-normalizer==3.4.2
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.10
packaging==25.0
pillow==11.2.1
python-dotenv==1.1.0
requests==2.32.3
tkcalendar==1.6.1
urllib3==2.4.0
